refactor(app): remove all Apollo link plugins and apollo-client utilities
Delete all Apollo link plugins (ApolloLinkPlugin, ApolloDynamicLink,
ConsoleLinkPlugin, NetworkErrorLinkPlugin, OmitTypenameLinkPlugin,
TenantHeaderLinkPlugin, AddQuerySelectionPlugin, ApolloCacheObjectIdPlugin)
and apollo-client cache utilities.

Move error overlay components from plugins/NetworkErrorLinkPlugin/ to
errors/overlay/ — still used by GraphQLClient's NetworkErrorPublishing.

Remove dead TenantHeaderLinkPlugin registration from TenancyProvider.
Remove AddGraphQLQuerySelection (unused component).
